Bug 392535 - Cell editor is now managed by the table
*) The active cell editor is now kept as an instance variable inside the
table istself. The table listens to a new layer event
(CellEditorCreatedEvent) in order to be notified about new editors.
*) All references to the ActiveCellEditorRegistry have been changed in
order to use the new API on the table itself.
*) The ActiveCellEditorRegistry have been marked as deprecated in order
to not break the backward compatibility.
*) The ScrollBarHandlerTemplate has been enhanced so that the table can
be set. This is required in order to properly be able to close the
editor when starting to scroll.

Change-Id: If99a1e9ae2f01f12a03fbf4424a5057ead70c77a
Signed-off-by: Michael Heiss <michael.heiss@salomon.at>