Bug 500788 - NatExporter always opens the exported result
An additional flag can be used to prevent the automatic opening
of the export result.

It would be possible initializing NatExporter without a shell,
but this approach crashes with the default IOutputStreamProvider.

The proposal is a new flag to let the user decide if the result should
be opend or not (may be in the UI of the application).

The default behaviour is not broken.

Change-Id: Icc5c3816af5e44cf15ad25f61b1176e3b0f3450a
Signed-off-by: Uwe Peuker <dev@upeuker.net>