Bug 379898 - Hierarchical tree table
Fixed issues in correlation with freeze handling. The
FreezeSelectionStrategy assumed that the FreezeLayer is built on top of
the SelectionLayer and therefore no position transformation was
performed. In case a transformation layer is added in between, the
freeze coordinates are incorrect, as the needed transformation is
missing. This for example came up when using the newly introduced
HierarchicalTreeLayer which is placed between the SelectionLayer and the
ViewportLayer, and therefore the FreezeLayer is created on top of the
HierarchicalTreeLayer. Therefore a position conversion is needed if the
scrollable layer in the ViewportLayer is not the SelectionLayer.

Additionally the index-position-transformation via LayerUtil needed to
be extended to handle added columns. As columns added by a layer have no
index, there needs to be a mechanism that allows to back convert such a
column. Therefore a modifier was introduced which allows to mark such a
column index.

Change-Id: I4d9896e3c8546975a72656f37188ddd64bbce048
Signed-off-by: Dirk Fauth <dirk.fauth@googlemail.com>