Bug 534446 - SelectAllCommand does not trigger repaint
When using the ResizeColumnHideShowLayer the CellSelectionEvent for
position 0/0 is not working when the first column is hidden. The column
position 0 is still in the layer stack, but it is hidden with a width of
0. Therefore the event is not processed to the NatTable level, as the
ViewportLayer will not process it further up the layer stack. Therefore
the position coordinates for the select all event needs to be evaluated
according to the column/row states.

Change-Id: I6d9e298757ab47c97d5c170fcd3b18db92e5fe4b
Signed-off-by: Dirk Fauth <dirk.fauth@googlemail.com>