Bug 567879 - Support separator in PopupContributionItem
PopupContributionItem has to override IContributionItem#isSeparator and
return true for separator items created with
PopupMenuBuilder#withSeperator(). Stored a separator flag as field in
PopupContributionItem and provided and additional constructor.
The logic of withMenuItemProvider() has been inlined to
withSeparator(String) to avoid the need to have a new API method to
create PopupContributionItem's via the additional constructor.

Change-Id: Ib484e320f73f712c46dcd9ca335a0c0430e36ccb